feat: add ScanNetworkIcon component
