+++ src/client/components/app/index.ts
@@ -9,6 +9,7 @@ export * from './Navigation';
 export * from './NoWalletCard';
 export * from './RecommendationsCard';
 export * from './Settings';
+export * from './ScanNetworkIcon';
 export * from './SummaryCard';
 export * from './TokenIcon';
 export * from './Transactions';
